randgen/mutations: move mutators to randgen
This commit moves random SQL statement mutators from the `mutations` to
the `randgen` package, which is a more appropriate home for them.

Note that there are still a few test utility functions in the
`mutations` package, though I'm not convinced they belong there. If
someone knows of a better package for them, I'd gladly eliminate the
`mutations` package altogether.

Release note: None
