+++ pkg/BUILD.bazel
@@ -202,7 +202,6 @@ ALL_TESTS = [
     "//pkg/sql/lex:lex_test",
     "//pkg/sql/lexbase:lexbase_test",
     "//pkg/sql/logictest:logictest_test",
-    "//pkg/sql/mutations:mutations_test",
     "//pkg/sql/oidext:oidext_test",
     "//pkg/sql/opt/bench:bench_test",
     "//pkg/sql/opt/cat:cat_test",
@@ -238,6 +237,7 @@ ALL_TESTS = [
     "//pkg/sql/privilege:privilege_test",
     "//pkg/sql/protoreflect:protoreflect_test",
     "//pkg/sql/querycache:querycache_test",
+    "//pkg/sql/randgen:randgen_test",
     "//pkg/sql/row:row_test",
     "//pkg/sql/rowcontainer:rowcontainer_test",
     "//pkg/sql/rowenc:rowenc_test",
