+++ pkg/cmd/smithcmp/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
     deps = [
         "//pkg/cmd/cmpconn",
         "//pkg/internal/sqlsmith",
-        "//pkg/sql/mutations",
         "//pkg/sql/randgen",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
