+++ pkg/internal/sqlsmith/BUILD.bazel
@@ -21,7 +21,6 @@ go_library(
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/typedesc",
-        "//pkg/sql/mutations",
         "//pkg/sql/parser",
         "//pkg/sql/randgen",
         "//pkg/sql/sem/builtins",
