+++ pkg/sql/logictest/BUILD.bazel
@@ -23,10 +23,10 @@ go_library(
         "//pkg/settings/cluster",
         "//pkg/sql",
         "//pkg/sql/execinfra",
-        "//pkg/sql/mutations",
         "//pkg/sql/parser",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
+        "//pkg/sql/randgen",
         "//pkg/sql/row",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
